fix(install): refuse process-group signals when wait ownership is unknown
Round-5 review (01KXAQXAFKV5ZCC3K7E2Q2HP1C) found that waitid failures
(ECHILD / auto-reaped child status) returned ProbeStopReason::WaitError,
but version_output_with_limits still called finish(), which treated every
non-LeaderExited stop as an owned live group and sent SIGTERM/SIGKILL to
the cached negative PGID. After the child status is consumed, that PID/PGID
can be reused — redirecting signals to an unrelated process group of the
same user (AGENTS Rust 015/018; same contract as csa_process WaitStateUnknown).

- Split OwnershipUnknown (no signals) from OutputStatError (still clean)
- Mark ownership_unknown on waitid/try_wait failure so Drop cannot escalate
- abandon_without_signaling() for exact-child try_wait only; never -pgid
- Extract install_provenance_probe.rs to stay under the monolith token gate
- Deterministic regressions: finish/Drop ownership-unknown leave hang alive;
  OutputStatError still kills; waitid ECHILD injection path does not signal

### AI Reviewer Metadata
- **Design Intent**: Match daemon_cleanup WaitStateUnknown — once wait
  ownership is unknown, never emit PID or negative-PGID signals from the
  version probe.
- **Key Decisions**: Separate wait-state loss from output-stat failure so
  owned leaders still get controlled cleanup; set the ownership flag at the
  waitid failure site so panic paths between poll and finish stay safe.
- **Reviewer Guidance**:
  - **Timing/Race Scenarios**: ECHILD after SIGCHLD auto-reap; Drop after
    mark_ownership_unknown without finish; OutputStatError with live group.
  - **Boundary Cases**: finish(OwnershipUnknown) returns cannot-safely-signal;
    version_output_with_limits short-circuits before finish on that path.
  - **Regression Tests Added**:
    version_probe_ownership_unknown_emits_no_pid_or_pgid_signal;
    version_probe_drop_after_ownership_unknown_does_not_signal;
    version_probe_output_stat_error_still_cleans_owned_group;
    version_probe_waitid_echild_does_not_signal_process_group.
